QuadTreeNode -> RootQuadTreeNode & StatusQuadTreeNode Tiles generation for the RawMandelbrotData
git-svn-id: http://uuh116.googlecode.com/svn/Nebulabrot@229 333f4dc9-c705-684f-d081-88386eedeb15

+++ src/net/lab0/nebula/color/PointValues.java
@@ -2,7 +2,7 @@ package net.lab0.nebula.color;
 
 public class PointValues
 {
-    public int minIter;
-    public int maxIter;
-    public int value;
+    public long minIter;
+    public long maxIter;
+    public long value;
 }
